suppress warnings
git-svn-id: https://svn.salilab.org/imp/trunk@17037 e91f8c7b-0e32-4433-a9ac-4502d39a4eab

+++ applications/foxs/lib/Gnuplot.cpp
@@ -73,7 +73,8 @@ void Gnuplot::print_canvas_script(const std::vector<std::string>& pdbs,
     std::string profile_file_name = pdbs[i] + ".dat";
     plt_file <<  "'" << profile_file_name << "' u 1:2 thru log(y) "
              << "w lines lw 2.5 lc rgb '#" << hex_color << "'";
-    if(i==pdbs.size()-1 || i==max_num-1) plt_file << std::endl;
+    if(i== static_cast<int>(pdbs.size())-1
+       || i==max_num-1) plt_file << std::endl;
     else plt_file << ",";
   }
   plt_file << "reset\n";
